Deploy monitoring when created Instead of deploying monitorings with a new deployment run, deploy it when created on MonitoringController.
