+++ projects/api/deployments/runs/runs.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """Runs API Router."""
-from fastapi import APIRouter, BackgroundTasks, Depends
+from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 
 from projects.controllers import DeploymentController, ProjectController
@@ -36,15 +36,13 @@ async def handle_list_runs(project_id: str,
     deployment_controller.raise_if_deployment_does_not_exist(deployment_id)
 
     run_controller = RunController(session)
-    runs = run_controller.list_runs(project_id=project_id,
-                                    deployment_id=deployment_id)
+    runs = run_controller.list_runs(deployment_id=deployment_id)
     return runs
 
 
 @router.post("")
 async def handle_post_runs(project_id: str,
                            deployment_id: str,
-                           background_tasks: BackgroundTasks,
                            session: Session = Depends(session_scope)):
     """
     Handles POST requests to /.
@@ -65,9 +63,8 @@ async def handle_post_runs(project_id: str,
     deployment_controller = DeploymentController(session)
     deployment_controller.raise_if_deployment_does_not_exist(deployment_id)
 
-    run_controller = RunController(session, background_tasks)
-    run = run_controller.create_run(project_id=project_id,
-                                    deployment_id=deployment_id)
+    run_controller = RunController(session)
+    run = run_controller.create_run(deployment_id=deployment_id)
     return run
 
 
@@ -97,9 +94,7 @@ async def handle_get_run(project_id: str,
     deployment_controller.raise_if_deployment_does_not_exist(deployment_id)
 
     run_controller = RunController(session)
-    run = run_controller.get_run(project_id=project_id,
-                                 deployment_id=deployment_id,
-                                 run_id=run_id)
+    run = run_controller.get_run(deployment_id=deployment_id)
     return run
