+++ projects/api/monitorings/monitorings.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """Monitorings API Router."""
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, BackgroundTasks, Depends
 from sqlalchemy.orm import Session
 
 import projects.schemas.monitoring
@@ -37,8 +37,7 @@ async def handle_list_monitorings(project_id: str,
     deployment_controller.raise_if_deployment_does_not_exist(deployment_id)
 
     monitoring_controller = MonitoringController(session)
-    monitorings = monitoring_controller.list_monitorings(project_id=project_id,
-                                                         deployment_id=deployment_id)
+    monitorings = monitoring_controller.list_monitorings(deployment_id=deployment_id)
     return monitorings
 
 
@@ -46,6 +45,7 @@ async def handle_list_monitorings(project_id: str,
 async def handle_post_monitorings(project_id: str,
                                   deployment_id: str,
                                   monitoring: projects.schemas.monitoring.MonitoringCreate,
+                                  background_tasks: BackgroundTasks,
                                   session: Session = Depends(session_scope)):
     """
     Handles POST requests to /.
@@ -67,9 +67,8 @@ async def handle_post_monitorings(project_id: str,
     deployment_controller = DeploymentController(session)
     deployment_controller.raise_if_deployment_does_not_exist(deployment_id)
 
-    monitoring_controller = MonitoringController(session)
-    monitoring = monitoring_controller.create_monitoring(project_id=project_id,
-                                                         deployment_id=deployment_id,
+    monitoring_controller = MonitoringController(session, background_tasks)
+    monitoring = monitoring_controller.create_monitoring(deployment_id=deployment_id,
                                                          monitoring=monitoring)
     return monitoring
 
@@ -100,7 +99,5 @@ async def handle_delete_monitorings(project_id: str,
     deployment_controller.raise_if_deployment_does_not_exist(deployment_id)
 
     monitoring_controller = MonitoringController(session)
-    response = monitoring_controller.delete_monitoring(uuid=monitoring_id,
-                                                       project_id=project_id,
-                                                       deployment_id=deployment_id)
+    response = monitoring_controller.delete_monitoring(uuid=monitoring_id)
     return response
