+++ projects/controllers/monitorings/monitorings.py
@@ -1,17 +1,21 @@
 # -*- coding: utf-8 -*-
 """Monitorings controller."""
 from projects import models, schemas
+from projects.controllers.deployments.runs.runs import RunController
 from projects.controllers.tasks import TaskController
 from projects.controllers.utils import uuid_alpha
 from projects.exceptions import NotFound
+from projects.kfp.monitorings import deploy_monitoring
 
 
 NOT_FOUND = NotFound("The specified monitoring does not exist")
 
 
 class MonitoringController:
-    def __init__(self, session):
+    def __init__(self, session, background_tasks=None):
         self.session = session
+        self.background_tasks = background_tasks
+        self.run_controller = RunController(session)
         self.task_controller = TaskController(session)
 
     def raise_if_monitoring_does_not_exist(self, monitoring_id: str):
@@ -33,13 +37,12 @@ class MonitoringController:
         if not exists:
             raise NotFound("The specified monitoring does not exist")
 
-    def list_monitorings(self, project_id: str, deployment_id: str):
+    def list_monitorings(self, deployment_id: str):
         """
         Lists all monitorings under a deployment.
 
         Parameters
         ----------
-        project_id : str
         deployment_id : str
 
         Returns
@@ -53,7 +56,7 @@ class MonitoringController:
 
         return schemas.MonitoringList.from_orm(monitorings, len(monitorings))
 
-    def create_monitoring(self, monitoring: schemas.MonitoringCreate, project_id: str, deployment_id: str):
+    def create_monitoring(self, monitoring: schemas.MonitoringCreate, deployment_id: str):
         """
         Creates a new monitoring in our database.
 
@@ -78,17 +81,28 @@ class MonitoringController:
         self.session.commit()
         self.session.refresh(monitoring)
 
+        # Deploy the new monitoring
+        deployment = self.session.query(models.Deployment).get(deployment_id)
+        run = self.run_controller.get_run(deployment_id)
+
+        self.background_task.add_task(
+            deploy_monitoring,
+            deployment_id=deployment_id,
+            experiment_id=deployment.experiment_id,
+            run_id=run.uuid,
+            task_id=monitoring.task_id,
+            monitoring_id=monitoring.uuid
+        )
+
         return schemas.Monitoring.from_orm(monitoring)
 
-    def delete_monitoring(self, uuid, project_id, deployment_id):
+    def delete_monitoring(self, uuid):
         """
         Delete a monitoring in our database.
 
         Parameters
         ----------
         uuid : str
-        project_id : str
-        deployment_id : str
 
         Returns
         -------
